api_video/requirements.txt
fastapi
mysqlclient
SQLAlchemy
python-decouple
uvicorn
celery==4.4.7
redis
flower
minio
Pillow
decord
requests-async
python-multipart